Upgrade to latest wasmparser/wasmencoder.

+++ src/ir/module.rs
@@ -43,8 +43,8 @@ pub struct MemorySegment {
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct TableData {
     pub ty: Type,
-    pub initial: u32,
-    pub max: Option<u32>,
+    pub initial: u64,
+    pub max: Option<u64>,
     pub func_elements: Option<Vec<Func>>,
 }
 
@@ -171,7 +171,7 @@ impl<'a> Module<'a> {
 }
 
 impl<'a> Module<'a> {
-    pub(crate) fn frontend_add_table(&mut self, ty: Type, initial: u32, max: Option<u32>) -> Table {
+    pub(crate) fn frontend_add_table(&mut self, ty: Type, initial: u64, max: Option<u64>) -> Table {
         let func_elements = Some(vec![]);
         self.tables.push(TableData {
             ty,
